Update README.md
To accelerate the creation of the New Garden of Eden, a paradigm of abundant living spaces and Free Private Cities, anchored in BTC Maximalist Web3 technology, AI, and ethical psycho-technologies like Yoga and Meditation. The objective is to offer an environment immune to corruption and coercion, where every individual can achieve eudaimonia, flourishing mentally, financially, and geographically.